+++ src/app/dashboard/page.tsx
@@ -1,5 +1,18 @@
-export default function DashboardPage() {
+import dynamic from 'next/dynamic'
+const Dashboard = dynamic(() => import('@/features/Dashboard/Dashboard').then(m => m.Dashboard), { ssr: false })
+import { getAccessToken, getUserInfo } from "@/shared/lib/auth.lib";
+import { LoginRequiredModal } from '@/shared/ui/organisms/LoginRequiredModal';
+
+export default async function DashboardPage() {
+  const [ accessToken, userInfo ] = await Promise.all([
+    getAccessToken(),
+    getUserInfo(),
+  ])
+
   return (
-    <h1 className="text-3xl font-bold text-center">Dashboard</h1>
+    <>
+      <LoginRequiredModal accessToken={accessToken} />
+      <Dashboard userInfo={userInfo} />
+    </>
   )
 }