+++ src/shared/types/global.types.ts
@@ -1,3 +1,14 @@
+import { AxiosError, AxiosResponse } from "axios";
+
+export type ThemeMode = 'light' | 'dark';
+
+export type ErrorCatched = {
+  message: string;
+  cause?: {
+    code: string
+  }
+}
+
 export type GeneralError = {
   response: {
     data: {
@@ -8,7 +19,15 @@ export type GeneralError = {
   }
 }
 
+export interface GeneralApiError extends Omit<AxiosError, 'response'> {
+  response: AxiosResponse<{
+    message: string;
+  }>;
+}
+
 export type CookieObject = {
   name: string;
   value: string;
 };
+
+export type UserRoles = 'user' | 'admin'