+++ src/shared/ui/atoms/PostalCodeInput.tsx
@@ -1,0 +1,5 @@
+export const PostalCodeInput = () => {
+  return (
+    <input type="text" inputMode="numeric" />
+  )
+}